Store Database Creation
I've created a database for a bike store. The bike store has 15 items for sale, and I've included the item ID, item name, price for sale, how many items are in inventory, and the amount of each item sold. I then ordered the items by price and found the total number of items sold from the store. This was done using SQL.